Make source positions 32-bit to reduce memory footprint
Change `Location` and `Span` fields from `usize` to `u32`, limiting
source size to 4 GiB and shrinking per-token and per-scope structures.
This reduces `Diagnostic` from 200 to 112 bytes and parser state from
896 to 544 bytes.

Also refactor the lexer to construct diagnostics on demand via
`failureDiagnostic()` instead of carrying a `Diagnostic` by value in the
result union, reducing `Lexer.Result` from 208 to 24 bytes.

Remove the `source_offset_overflow` storage failure variant since
sources are now checked against the 32-bit limit before scanning, and
update documentation to reflect the new size guarantees and performance
gains.

+++ src/syntax_event.zig
@@ -422,7 +422,7 @@ test "recording sink preserves event order and borrowed spans" {
     var sink: RecordingSink(8) = .{};
 
     const span = struct {
-        fn at(offset: usize, len: usize) location.Span {
+        fn at(offset: u32, len: u32) location.Span {
             return .{
                 .start = .{ .byte_offset = offset, .line = 1, .byte_column = offset + 1 },
                 .byte_len = len,
